pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-tri7b"
version = "0.1.0"
description = "RAG pipeline with LangChain + LangGraph (local models only)"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Team", email = "team@example.com" }
]
dependencies = [
  "torch",
  "torchvision",
  "torchaudio",
  "transformers",
  "tokenizers",
  "sentence-transformers",
  "numpy",
  "langchain",
  "langchain-community",
  "langchain-core",
  "langchain-huggingface",
  "langgraph",
  "faiss-cpu",
  "pypdf",
  "pyhwp",
  "python-docx",
  "huggingface-hub",
  "accelerate",
  "onnxruntime",
  "soundfile",
  "g2pkk",
  "jamo",
  "anyascii",
  "bitsandbytes",
  "vllm>=0.6.4",
  "six"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
